### 2023-04-17 #5378 @editorialbot
[PRE REVIEW]: SleepECG: a Python package for sleep staging based on heart rate
**Submitting author:** @cbrnr (<a href="http://orcid.org/0000-0002-6030-2233">Clemens Brunner</a>)
**Repository:** https://github.com/cbrnr/sleepecg
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.4
