+++ setup.cfg
@@ -33,7 +33,7 @@ install_requires =
     tqdm>=4.60.0
 
 [options.extras_require]
-full =
+full =  # complete package functionality
     joblib>=1.0.0
     matplotlib>=3.5.0
     mne>=1.0.0
@@ -41,7 +41,7 @@ full =
     tensorflow>=2.7.0;python_version<'3.11'
     wfdb>=3.4.0
 
-dev =
+dev =  # everything needed for development
     black>=22.6.0
     flake8>=4.0.0
     joblib>=1.0.0
@@ -59,11 +59,11 @@ dev =
     tensorflow>=2.7.0;python_version<'3.11'
     wfdb>=3.4.0
 
-doc =
+doc =  # RTD uses this when building the documentation
     mkdocs-material>=8.4.0
     mkdocstrings-python>=0.8.2
 
-cibw =
+cibw =  # cibuildwheel uses this for running the test suite
     mne>=1.0.0
     numba>=0.55.0;python_version<'3.11'
     pyEDFlib>=0.1.25
