### 2023-04-19 #5378 @emdupre
Could you please also update your [associated BibTeX file](https://github.com/cbrnr/sleepecg/blob/joss/joss/paper.bib) ? Currently compilation is returning the following error:

```
Failed to parse BibTeX on value "--266" (NAME) ["@", #<BibTeX::Entry >, {:author=>["Agnew Jr., H. W. and Webb, W. B. Webb and Williams, R. L."], :title=>["{The first night effect: an EEG study of sleep}"], :journal=>["Psychophysiology"], :volume=>"2", :number=>"3", :year=>"1966", :pages=>"263"}]
```

You can fix this by enclosing the ranges in all `pages` field entries in curly brackets or double-quotes.
