### 2023-04-19 f77e73b
Fix warnings during build time (#144)
* Explicitly include classifiers and test

* Reformat setup.cfg

* Add docstring

* Also manually include test

* test should be a package

* Add docstring

* Move non-setuptools config to other files

* Go back to single quotes

* Move pydocstyle config to separate file

* Bump versions

* Use correct package name

* Remove comment

* Remove another comment

* Fix black

* Bump actions

* Fix actions

* Try test directory

* Try project

* Next try with folder

* Back to package

* See if test package is the problem

* Revert including classifiers

* Use 
[…]
