+++ .github/workflows/test.yml
@@ -40,7 +40,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Build wheels and run pytest
-        uses: pypa/cibuildwheel@v2.11.2
+        uses: pypa/cibuildwheel@v2.12.1
 
       - uses: actions/upload-artifact@v3
         with:
