+++ setup.cfg
@@ -21,10 +21,9 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 keywords = sleep, ecg, qrs, peak
 
-
 [options]
 include_package_data = True
-packages = find:
+packages = find_namespace:
 python_requires = >=3.8
 install_requires =
     numpy>=1.20.0
@@ -33,10 +32,8 @@ install_requires =
     scipy>=1.7.0
     tqdm>=4.60.0
 
-
 [options.extras_require]
-# complete package functionality
-full =
+full =  # complete package functionality
     joblib>=1.0.0
     matplotlib>=3.5.0
     mne>=1.0.0
@@ -44,8 +41,7 @@ full =
     tensorflow>=2.7.0;python_version<'3.11'
     wfdb>=3.4.0
 
-# everything needed for development
-dev =
+dev =  # everything needed for development
     black>=22.6.0
     flake8>=4.0.0
     joblib>=1.0.0
@@ -63,43 +59,19 @@ dev =
     tensorflow>=2.7.0;python_version<'3.11'
     wfdb>=3.4.0
 
-# Read the Docs uses this when building the documentation
-doc =
+doc =  # RTD uses this when building the documentation
     mkdocs-material>=8.4.0
     mkdocstrings-python>=0.8.2
 
-# cibuildwheel uses this for running the test suite
-# pytest is installed automatically (see pyproject.toml[tool.cibuildwheel])
-cibw =
+cibw =  # cibuildwheel uses this for running the test suite
     mne>=1.0.0
     numba>=0.55.0;python_version<'3.11'
     pyEDFlib>=0.1.25
     wfdb>=3.4.0
 
-
 [options.package_data]
 * = *.pyi
-
 
 [options.packages.find]
 exclude =
     examples
-
-
-[flake8]
-extend-ignore = E203  # for Black compatibility
-max-line-length = 92
-max-doc-length = 92
-exclude =
-    ./build/*
-per-file-ignores = __init__.py:F401
-
-
-[pydocstyle]
-convention = numpy
-add_ignore = D105  # missing docstring in magic method
-
-
-[tool:pytest]
-markers =
-    c_extension: mark a test as belonging to a C extension
