### 2023-04-24 #5378 @tompollard
Very cool, first time that I've received an invitation like this through GitHub!

I am a little overcommitted, but since this is something new and interesting I'd be happy to provide a review (from a quick look over the repo, it looks like a very clean, well-managed codebase).

I would also recommend Benjamin @bemoody as a reviewer. He is overcommitted like everyone else, but interest value might override this?
