### 2023-04-25 #5378 @sangfrois
Hi  @emdupre !
So cool to hear from you through this platform! I greatly appreciate the invite. Though, I am also on a short bandwidth lately.

I'M FINALLY SENDING MY MASTERS THESIS EVERYBODY 💯 🥇  lol
... after practically four years.  Never too late!

Anyways, I will be wrapping up a preprint in the coming weeks so I don't think I can be of much help for reviewing in time.

Please do hit me up if ever there are opportunities to review papers akin to this one (the work looks pretty neat).
I will be up for the task at another time, promise!

hope to see you at OHBM, including you too @DominiqueMakowski
