### 2023-04-25 #5378 @peterakirk
Hi @emdupre, I don't have expertise in sleep stage assessment, but I'd be happy to have a look through to check the documentation, installation, and a quick benchmarking to make sure HR/HRV metrics seem appropriate.
