### 2023-05-03 ac03d63
Add default in docstring
