+++ sleepecg/feature_extraction.py
@@ -4,8 +4,6 @@
 
 """Functions and utilities related to feature extraction."""
 
-from __future__ import annotations
-
 import warnings
 from typing import Iterable, Optional
 
