+++ sleepecg/io/gudb.py
@@ -4,8 +4,6 @@
 
 """Utilities for downloading GUDB data."""
 
-from __future__ import annotations
-
 from pathlib import Path
 from typing import Optional
 
