+++ sleepecg/io/ecg_readers.py
@@ -4,8 +4,6 @@
 
 """Functions for reading datasets containing ECG and beat annotations."""
 
-from __future__ import annotations
-
 from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Iterator, Optional
