+++ examples/benchmark/utils.py
@@ -4,8 +4,6 @@
 
 """Utilities for runtime and detection quality benchmarks."""
 
-from __future__ import annotations
-
 import time
 from typing import Any, Iterator
 
