+++ sleepecg/config.py
@@ -4,8 +4,6 @@
 
 """Functions for getting and setting configuration values."""
 
-from __future__ import annotations
-
 from pathlib import Path
 from typing import Any, Optional
 
