+++ sleepecg/classification.py
@@ -4,8 +4,6 @@
 
 """Functions related to classifier training and evaluation."""
 
-from __future__ import annotations
-
 import shutil
 from dataclasses import dataclass
 from pathlib import Path
