+++ sleepecg/test/test_sleep_readers.py
@@ -4,8 +4,6 @@
 
 """Tests for sleep data reader functions."""
 
-from __future__ import annotations
-
 import datetime
 from pathlib import Path
 
