+++ setup.cfg
@@ -9,7 +9,6 @@ maintainer_email = clemens.brunner@gmail.com
 classifiers =
     License :: OSI Approved :: BSD License
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
@@ -24,7 +23,7 @@ keywords = sleep, ecg, qrs, peak
 [options]
 include_package_data = True
 packages = find_namespace:
-python_requires = >=3.8
+python_requires = >=3.9
 install_requires =
     numpy>=1.20.0
     PyYAML>=5.4.0
