+++ setup.cfg
@@ -9,7 +9,6 @@ maintainer_email = clemens.brunner@gmail.com
 classifiers =
     License :: OSI Approved :: BSD License
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: Implementation :: CPython
@@ -23,7 +22,7 @@ keywords = sleep, ecg, qrs, peak
 [options]
 include_package_data = True
 packages = find_namespace:
-python_requires = >=3.9
+python_requires = >=3.10
 install_requires =
     numpy>=1.20.0
     PyYAML>=5.4.0
