+++ readthedocs.yml
@@ -4,9 +4,9 @@
 version: 2
 
 build:
-  os: ubuntu-20.04
+  os: ubuntu-22.04
   tools:
-    python: "3.9"
+    python: "3.11"
 
 mkdocs:
   configuration: mkdocs.yml
