+++ setup.cfg
@@ -40,6 +40,7 @@ full =  # complete package functionality
     numba>=0.55.0
     tensorflow>=2.7.0; sys_platform != 'darwin'
     tensorflow-macos>=2.7.0; sys_platform == 'darwin' and python_version < '3.11'
+    tensorflow-metal>=0.8.0; sys_platform == 'darwin' and python_version < '3.11'
     wfdb>=3.4.0
 
 dev =  # everything needed for development
@@ -59,6 +60,7 @@ dev =  # everything needed for development
     sphinx>=4.1.0
     tensorflow>=2.7.0; sys_platform != 'darwin'
     tensorflow-macos>=2.7.0; sys_platform == 'darwin' and python_version < '3.11'
+    tensorflow-metal>=0.8.0; sys_platform == 'darwin' and python_version < '3.11'
     wfdb>=3.4.0
 
 doc =  # RTD uses this when building the documentation
