### 2023-05-04 3b8849e
Transform warnings into errors
