+++ pyproject.toml
@@ -36,6 +36,7 @@ add_ignore = 'D105'
 [tool.pytest.ini_options]
 markers = ['c_extension']
 filterwarnings = [
+    'error',
     'ignore:.*sample_rate:DeprecationWarning:pyedflib',
     'ignore:.*pkg_resources:DeprecationWarning',
 ]
