+++ sleepecg/test/test_sleep_readers.py
@@ -10,7 +10,11 @@ import datetime
 from pathlib import Path
 
 import numpy as np
-import scipy.misc
+
+try:
+    from scipy.datasets import electrocardiogram  # SciPy ≥ 1.10
+except ImportError:
+    from scipy.misc import electrocardiogram  # SciPy < 1.10
 from pyedflib import highlevel
 
 from sleepecg import SleepStage, read_mesa, read_shhs, read_slpdb
@@ -19,7 +23,7 @@ from sleepecg.io.sleep_readers import Gender
 
 def _dummy_nsrr_edf(filename: str, hours: float, ecg_channel: str):
     ECG_FS = 360
-    ecg_5_min = scipy.misc.electrocardiogram()
+    ecg_5_min = electrocardiogram()
     seconds = int(hours * 60 * 60)
     ecg = np.tile(ecg_5_min, int(np.ceil(seconds / 300)))[np.newaxis, : seconds * ECG_FS]
     signal_headers = highlevel.make_signal_headers([ecg_channel], sample_frequency=ECG_FS)
