+++ pyproject.toml
@@ -35,3 +35,8 @@ add_ignore = 'D105'
 
 [tool.pytest.ini_options]
 markers = ['c_extension']
+filterwarnings = [
+    'error',
+    'ignore:.*sample_rate:DeprecationWarning:pyedflib',
+    'ignore:.*pkg_resources:DeprecationWarning',
+]
