### 2023-05-06 4eece2b
test detect_heartbeats with numba backend on python 3.11
