### 2023-05-06 e029bfb
Silence TensorFlow messages (#146)
* Silence TensorFlow messages

* Fix type error

* Fix Black

* Fix Black again

* Add default in docstring
