### 2023-05-11 0258ac3
Replace setuptools.cfg with pyproject.toml
