+++ setup.cfg
@@ -1,81 +1,0 @@
-[metadata]
-name = sleepecg
-version = attr: sleepecg.__version__
-url = https://github.com/cbrnr/sleepecg
-author = Florian Hofer
-author_email = hofaflo@gmail.com
-maintainer = Clemens Brunner
-maintainer_email = clemens.brunner@gmail.com
-classifiers =
-    License :: OSI Approved :: BSD License
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: Implementation :: CPython
-    Typing :: Typed
-license = BSD 3-Clause License
-description = A package for sleep stage classification using ECG data
-long_description = file: README.md
-long_description_content_type = text/markdown
-keywords = sleep, ecg, qrs, peak
-
-[options]
-include_package_data = True
-packages = find_namespace:
-python_requires = >=3.8
-install_requires =
-    numpy>=1.20.0
-    PyYAML>=5.4.0
-    requests>=2.25.0
-    scipy>=1.7.0
-    tqdm>=4.60.0
-
-[options.extras_require]
-full =  # complete package functionality
-    joblib>=1.0.0
-    matplotlib>=3.5.0
-    mne>=1.0.0
-    numba>=0.55.0
-    tensorflow>=2.7.0; sys_platform != 'darwin'
-    tensorflow-macos>=2.7.0; sys_platform == 'darwin' and python_version < '3.11'
-    tensorflow-metal>=0.8.0; sys_platform == 'darwin' and python_version < '3.11'
-    wfdb>=3.4.0
-
-dev =  # everything needed for development
-    black>=22.6.0
-    flake8>=4.0.0
-    joblib>=1.0.0
-    matplotlib>=3.5.0
-    mkdocs-material>=8.4.0
-    mkdocstrings-python>=0.7.1
-    mne>=1.0.0
-    mypy>=0.991
-    numba>=0.55.0
-    pre-commit>=2.13.0
-    pyEDFlib>=0.1.25
-    pytest>=6.2.0
-    setuptools>=56.0.0
-    sphinx>=4.1.0
-    tensorflow>=2.7.0; sys_platform != 'darwin'
-    tensorflow-macos>=2.7.0; sys_platform == 'darwin' and python_version < '3.11'
-    tensorflow-metal>=0.8.0; sys_platform == 'darwin' and python_version < '3.11'
-    wfdb>=3.4.0
-
-doc =  # RTD uses this when building the documentation
-    mkdocs-material>=8.4.0
-    mkdocstrings-python>=0.8.2
-
-cibw =  # cibuildwheel uses this for running the test suite
-    mne>=1.0.0
-    numba>=0.55.0
-    pyEDFlib>=0.1.25
-    wfdb>=3.4.0
-
-[options.package_data]
-* = *.pyi
-
-[options.packages.find]
-exclude =
-    examples
