### 2023-05-11 2950910
Replace setuptools.cfg with pyproject.toml (#154)
