### 2023-05-15 cbrnr/sleepecg#155 @richrobe
[JOSS Review] Information about "Movement Data"
(as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5411))

In the README, you wrote: "Based only on ECG (and to a lesser extent also movement data), `SleepECG` provides functions for [...]". I'm wondering where the movement data comes from since you didn't mention it elsewhere and also have no functions for importing, processing, and extracting features from acceleration or actigraphy data.
