### 2023-05-15 cbrnr/sleepecg#156 @richrobe
[JOSS Review] Paper Remarks
(as part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5411))

The paper is easy written, easy to follow, and nicely outlines the need of the `SleepECG` package. It provides a good motivation, especially regarding the reproducibility of the existing approaches. Just some minor comments:
- To better understand the scope of the package at first sight, an overview image showing the typical workflow for ECG-based sleep staging (data download, data cleaning, pre-processing, feature extraction, training, ...), and where `SleepECG` could help in this process (namely everywhere) would be beneficial.
- Could you add some more information on the pre-trained classifiers you used in your paper in terms of network architecture, training & evaluation process, etc.? Is there already a publication on it (if yes, could you link it?)
