### 2023-05-15 cbrnr/sleepecg#157 @richrobe
[JOSS Review] Installation fails on macOS
(as part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5411))

The straightforward installation of `SleepECG` fails on my system (Python 3.9 on 2017 Macbook Pro with macOS Ventura 13.3, Intel CPU, no GPU) using the current `pyproject.toml` version with the following error message:
with the following error message:
```
Could not find a version that satisfies the requirement tensorflow-macos>=2.7.0; (sys_platform == "darwin" and python_version < "3.11") and extra == "full" (from sleepecg[full]) (from versions: none)
```
Replacing the optional dependencies `tensorflow-macos` and `tensorflow-metal` by `tensorflow` in the `pyproject.toml` file fixes the issue and correctly installs the package with all dependencies.
