### 2023-05-15 cbrnr/sleepecg#158 @richrobe
[JOSS Review] Installing SleepECG with optional dependencies
(as part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5411))

When installing `SleepECG` with extras, following the instructions in the [Contributing](https://sleepecg.readthedocs.io/en/stable/#contributing) guide dev dependencies fails when using `zsh` as shell (which is now the default unser ) because square brackets need to be escaped. Thus, I would maybe suggest to add an additional note that zsh users should call ``pip install -e ".[dev]"`` (with quotes). This command should also work when using other shells (such as `bash`).
