### 2023-05-15 cbrnr/sleepecg#159 @richrobe
[JOSS Review] Add plt.show() to example script
(as part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5411))

The hypnogram in the example `try_ws_gru_mesa.py` isn't shown when running the code because `plt.show()` is missing in the end. Adding this line correctly displays the figure.
