### 2023-05-15 #5411 @richrobe
Hi @cbrnr,

thank you for putting together such a lovely package! In general, I like it very much and see it as a valuable contribution to sleep staging estimation from ECG data. After having gone through code, documentation and paper I'll summarize my thoughts below and have created extra issues in the package repository.

## General Remarks
- In the README, you wrote: "Based only on ECG (and to a lesser extent also movement data), `SleepECG` provides functions for [...]". I'm wondering where the movement data comes from since you didn't mention it elsewhere and also have no functions for importing, processing, and extracting features from acceleration or actigraphy data.

See also
- https://github.com/cbrnr/sleepecg/issues/155

## Paper
The paper is easy written, easy to follow, and nicely outlines the need of the `SleepECG` package. It provides a good motivation, especially regarding the reproducibility of the existing approaches. Just some minor comments:
- To better understand the scope of the package at first sight, an overview image showing the typical workflow for ECG-based sleep staging (data download, data cleaning, pre-processing, feature extraction, training, ...), and where `SleepECG` could help in this process (namely everywhere) would be beneficial.
- Could you add some more information on the pre-trained classifiers you used in your paper in terms of network architecture, training & evaluation process, etc.? Is there already a publication on it (if yes, could you link it?)

See also
- https://github.com/cbrnr/sleepecg/issues/156

## Installation
- The straightforward installation of `SleepECG` fails on my system (Python 3.9 on 2017 Macbook Pro with macOS Ventura 13.3, Intel CPU, no GPU) using the current `pyproject.toml` version with the following error message:
with the following error message:
```
Could not find a version that satisfies the requirement tensorflow-macos>=2.7.0; (sys_platform == "darwin" and python_version < "3.11") and extra == "full" (from sleepecg[full]) (from versions: none)
```
However, replacing the optional dependencies `tensorflow-macos` and `tensorflow-metal` by `tensorflow` in the `pyproject.toml` file fixes the issue and correctly installs the package with all dependencies.
- Additionally, when installing `SleepECG` with extras, following the instructions in the [Contributing](https://sleepecg.readthedocs.io/en/stable/#contributing) guide dev dependencies fails when using `zsh` as shell (which is now the default unser ) because square brackets need to be escaped. Thus, I would maybe suggest to add an additional note that zsh users should call ``pip install -e ".[dev]"`` (with quotes). This command should also work when using other shells (such as `bash`).

See also
- https://github.com/cbrnr/sleepecg/issues/157
- https://github.com/cbrnr/sleepecg/issues/158

## Examples
- Example `try_ws_gru_mesa.py`: hypnogram isn't shown when running the code because `plt.show()` is missing in the end. Adding this line correctly displays the figure.

See also
- https://github.com/cbrnr/sleepecg/issues/159

Other than that, I have nothing to add. Let me know what you think of my suggestions! 🙂
