### 2023-05-15 cbrnr/sleepecg#157 @cbrnr
That's weird, I get the same (or at least similar error) on my Intel Mac (I only tested on my M2 Mac previously). My idea was to provide optimized builds/packages for Macs, meaning native ARM64 support (`tensorflow-macos`) as well as Metal support (`tensorflow-macos`) for 3.8 <= Python <= 3.11. The "official" `tensorflow` package does not support ARM64.

Do you know which requirements exactly cannot be satisfied? A `tensorflow-macos` wheel should be available for your platform (https://pypi.org/project/tensorflow-macos/#files). I get this error with `tensorflow-metal` even though a matching package seems to be available.

If this cannot be resolved, I'm happy to revert to standard `tensorflow` (although it really bugs me that this still does not support ARM64, but there will be a wheel in the next release). Maybe we should have separate TF dependencies for each architecture?
