### 2023-05-15 cbrnr/sleepecg#155 @cbrnr
You're right, actigraphy data has been on our todo list ever since, but so far there is nothing in the package. I'll remove the part about movement data (we can add it again once we actually support it).
