### 2023-05-15 ee981ac
Escape square brackets (#160)
