### 2023-05-15 d3d15d5
Show plot in example (#162)
