### 2023-05-16 cbrnr/sleepecg#155 @richrobe
Alright! Let me know when you're interested in including actigraphy data as we're also working on sleep staging using wearable sensors (actigraphy, ECG, respiration, ...) and have some code lying around for feature extraction and classification from actigraphy data that you might be interested in using 🙂
