### 2023-05-16 cbrnr/sleepecg#157 @richrobe
I directly get an error when trying to manually install the mac-native tensorflow packages `tensorflow-mac` and `tensorflow-metal` via pip:

```
$ pip install tensorflow-macos
>>> ERROR: Could not find a version that satisfies the requirement tensorflow-macos (from versions: none)
>>> ERROR: No matching distribution found for tensorflow-macos
$ pip install tensorflow-metal
>>> ERROR: Could not find a version that satisfies the requirement tensorflow-metal (from versions: none)
>>> ERROR: No matching distribution found for tensorflow-metal
```

which is weird because, as you correctly pointed out, there **should** be a build available for my platform...
