### 2023-05-16 cbrnr/sleepecg#155 @cbrnr
This sounds great! We could discuss this anytime, it's just that I don't have a lot of time currently to actually implement anything as substantial as that.
