### 2023-05-16 cbrnr/sleepecg#157 @cbrnr
Strange. I think I tried installing just these packages yesterday, and it worked. Can you try and see if you can get more information out of pip with `-vvv`?
