### 2023-05-16 cbrnr/sleepecg#156 @cbrnr
Thanks, adding an overview figure is a great idea!

Regarding our pre-trained classifiers, we don't have a publication on that yet, so this will explained in all the details there. How much information do you think makes sense here? The scripts that we used to train are available inside [`examples/classifiers/`](https://github.com/cbrnr/sleepecg/tree/main/examples/classifiers), but I feel like this is way too much detail for this paper. What about including what features we used (all time-based and frequency-based HRV features, the recording starting time, age, and gender) as well as mention that we used a neural network with gated recurrent units (GRUs)? Would that be sufficient?
