### 2023-05-16 cbrnr/sleepecg#156 @richrobe
Yes, that sounds good! 🙂
