### 2023-05-16 cbrnr/sleepecg#156 @cbrnr
I've updated the paper, please [take a look](https://github.com/cbrnr/sleepecg/suites/12925945492/artifacts/698256375) and let me know if you would like me to change anything. If you're happy with the changes, feel free to close this issue.
