### 2023-05-16 cbrnr/sleepecg#157 @cbrnr
I can't really test this on my old Intel Mac, because I'm stuck with macOS 11.7.6. The problem is that the latest `tensorflow-macos` and `tensorflow-metal` versions require macOS 12, so unfortunately, I can't test this. I will check on my new ARM64 Mac though (I expect that everything works). Maybe the best solution is really to require `tensorflow` on Intel Macs and `tensorflow-macos`/`tensorflow-metal` on ARM64 Macs. I think it should be possible to specify these requirements in `pyproject.toml`.
