### 2023-05-18 #5411 @tompollard
Hi @cbrnr -

I agree with @richrobe. `sleepecg` is a helpful, well-described, easy to use software package. I added some thoughts below, but I also think that the software is in great shape as it stands. My comments mostly relate to ongoing maintenance tasks rather than things that need to be dealt with right now.

While I looked over the code (which is cleaner than 99% of code I come across, and certainly all of mine!) I didn't do a deep dive into any of the algorithms. As the package provides a framework for supporting a breadth of algorithms and does not claim to implement successful, validated sleep staging (at least right now), this didn't seem especially necessary.

# Installation

## Dependencies

For development I'd usually expect to see a file containing the dependencies somewhere (e.g. `requirements.txt` or `environment.yml`). Does this exist? I see that the dependencies are listed in `README.md`, but it would be nice to have a file that can be used by package managers (e.g. pip, conda).

## pip

`pip install sleepecg` correctly installs `sleepecg-0.5.4` on my machine (Apple M1 Max, macOS Ventura 13.3, Python 3.10, pip 23.0).

As mentioned by @richrobe in https://github.com/cbrnr/sleepecg/issues/158, `pip install sleepecg[full]` displays a `no matches found: sleepecg[full]` error in zshell.

`pip install 'sleepecg[full]'` correctly installs `sleepecg-0.5.4` (full) on my machine.

## import

When importing sleepecg (v0.5.4), I see the following deprecation notices:

```
>>> import sleepecg
/Users/tompollard/sand/sleepecg/env/lib/python3.10/site-packages/sleepecg/heartbeats.py:638: NumbaDeprecationWarning: The 'nopython' keyword argument was not supplied to the 'numba.jit' decorator. The implicit default value for this argument is currently False, but it will be changed to True in Numba 0.59.0. See https://numba.readthedocs.io/en/stable/reference/deprecation.html#deprecation-of-object-mode-fall-back-behaviour-when-using-jit for details.
  _squared_moving_integration_numba = jit(_squared_moving_integration_py)
/Users/tompollard/sand/sleepecg/env/lib/python3.10/site-packages/sleepecg/heartbeats.py:639: NumbaDeprecationWarning: The 'nopython' keyword argument was not supplied to the 'numba.jit' decorator. The implicit default value for this argument is currently False, but it will be changed to True in Numba 0.59.0. See https://numba.readthedocs.io/en/stable/reference/deprecation.html#deprecation-of-object-mode-fall-back-behaviour-when-using-jit for details.
  _thresholding_numba = jit(_thresholding_py)
```

It looks like this is already addressed in the repo, so I haven't raised an issue:
https://github.com/cbrnr/sleepecg/blob/d3d15d5bfa6386bdc68fb3c9bfc58d8da5f50d01/sleepecg/heartbeats.py#L638-L640

# Code

## Relative vs absolute imports

Absolute imports seem to be generally preferred by Python developers (https://peps.python.org/pep-0008/#imports). sleepecg uses relative imports in places (e.g. `from .physionet import _list_physionet`). Is this intentional? If not, perhaps switch to absolute imports at some point?

## Scripts

The scripts (e.g. `examples/benchmark/plot_benchmark_results.py`) are missing structure that I think can be helpful (e.g. `if __name__ == "__main__":`). Minor thing and not necessary, but might be worth adding some of this structure (e.g. to prevent them running when imported)? This is the outline I use: https://www.annasyme.com/docs/python_structure.html.

## Command line interface

Adding a command line interface for the package via `pyproject.toml` might be a nice addition, though not necessary.

## scipy.misc.electrocardiogram has been deprecated

The example code at https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage uses `scipy.misc.electrocardiogram` which has been deprecated. Needs updating to `scipy.datasets.electrocardiogram` at some point.

```
ecg = electrocardiogram()

<stdin>:1: DeprecationWarning: scipy.misc.electrocardiogram has been deprecated in SciPy v1.10.0; and will be completely removed in SciPy v1.12.0. Dataset methods have moved into the scipy.datasets module. Use scipy.datasets.electrocardiogram instead.
```

# Paper

Comments on the paper at: https://github.com/openjournals/joss-papers/blob/joss.05411/joss.05411/10.21105.joss.05411.pdf

## Figure 1

I'm not sure I understand the Y-axis labels on Figure 1. Could you add details to either the axis label or the caption? The caption of axis label should highlight the log scale. LTDB should be defined somewhere in the text or caption somewhere. The algorithms should also be defined (e.g. it is unclear to me what "mne" means).

## Add heartbeat detection metrics figure?

There is a metrics figure at https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage that I think would be a nice addition to the paper.

## Add sleep staging figure?

Given the focus of the package is sleep staging, I think it would be nice to add a figure demonstrating the sleep staging functionality. A (ideally non-cheat!) version of the hypnogram at https://sleepecg.readthedocs.io/en/stable/plot/ might be a good one to include?

## State of field

I'm not familiar with the literature on sleep-staging with ECG, but I do see there are one or two open source codebases that claim to enable ecg-based sleep staging. I wonder whether these should be briefly mentioned?

e.g. should comparisons be made to https://github.com/bdsp-core/ecg_respiration_sleep_staging? The repo doesn't look completely baked (e.g. "[training_code_still_messy](https://github.com/bdsp-core/ecg_respiration_sleep_staging/tree/master/training_code_still_messy))", so perhaps it reinforces your point that "either the processing pipeline or the dataset used to generate the published results are not publicly available."

## Reiterate point that sleep staging is not a solved problem?

I think it would be good to reiterate the point that sleep staging is not a solved problem. This is mentioned in the paper, but I think it is helpful for the user to be reminded that the package is not claiming (currently) to be able to successfully stage sleep with ECG.

## References

Not completely necessary, but ideally I think the algorithms and datasets should be formally cited in the paper if they are being discussed. This helps to credit the creators and also supports the idea of reproducibility.

# Archival version of the code

It wasn't clear to me whether there is an archival version of the software anywhere. If not, it would be great if could archive a copy (with DOI) somewhere like Zenodo or Figshare. I have a conflict of interest, but we would also be happy to host the software on [PhysioNet](https://physionet.org/).

# Documentation

`import numpy as np` should be moved to the code block above at https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage (to avoid `rri = 1000 * np.diff(beats) / fs` failing with an import error).
