### 2023-05-18 cbrnr/sleepecg#163 @tompollard
add requirements.txt to allow quick installation of dependencies?
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

For development I'd usually expect to see a file containing the dependencies somewhere (e.g. requirements.txt or environment.yml). Does this exist? I see that the dependencies are listed in README.md, but it would be nice to have a file that can be used by package managers (e.g. pip, conda).
