### 2023-05-18 cbrnr/sleepecg#164 @tompollard
switch from relative to absolute imports?
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

Absolute imports seem to be generally preferred by Python developers (https://peps.python.org/pep-0008/#imports). sleepecg uses relative imports in places (e.g. from .physionet import _list_physionet). Is this intentional? If not, perhaps switch to absolute imports at some point?
