### 2023-05-18 cbrnr/sleepecg#165 @tompollard
add structure to python scripts? (e.g. `if __name__ == "__main__":`)
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

The scripts (e.g. `examples/benchmark/plot_benchmark_results.py`) are missing structure that I think can be helpful (e.g. `if __name__ == "__main__":`).

Minor thing and not necessary, but might be worth adding some of this structure (e.g. to prevent them running when imported and for ease of running at the command line)?

This is the outline that I use: https://www.annasyme.com/docs/python_structure.html
