### 2023-05-18 cbrnr/sleepecg#166 @tompollard
`scipy.misc.electrocardiogram` has been deprecated
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

The example code at https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage (and elsewhere) uses `scipy.misc.electrocardiogram` which has been deprecated.

```
ecg = electrocardiogram()

<stdin>:1: DeprecationWarning: scipy.misc.electrocardiogram has been deprecated in SciPy v1.10.0; and will be completely removed in SciPy v1.12.0. Dataset methods have moved into the scipy.datasets module. Use scipy.datasets.electrocardiogram instead.
```

At some point the examples should be updated to use the new location: (`scipy.datasets.electrocardiogram`).
