### 2023-05-18 cbrnr/sleepecg#167 @tompollard
numpy referenced before it is imported in the `heartbeat_detection` documentation
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage has a nice overview of the heartbeat detection functionality.

`import numpy as np` should be moved to the code block above to avoid `rri = 1000 * np.diff(beats) / fs` failing with an import error.
