### 2023-05-18 cbrnr/sleepecg#168 @tompollard
Add figures to paper (heartbeat detection metrics; sleep staging)?
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

There is a metrics figure at https://sleepecg.readthedocs.io/en/stable/heartbeat_detection/#usage that I think would be a nice addition to the paper.

Given the focus of the package is sleep staging, it would also be nice to add a figure demonstrating the sleep staging functionality. A (ideally non-cheat!) version of the hypnogram at https://sleepecg.readthedocs.io/en/stable/plot/ might be a good one to include?

Neither of these are necessary, so please feel free to close this issue if you disagree or don't think it is worth the trouble!
