### 2023-05-18 cbrnr/sleepecg#169 @tompollard
Cite relevant codebases in JOSS paper
This is a quick issue relating to the JOSS review comments at: https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1548163405

I am not familiar with the literature on sleep-staging with ECG, but I do see there are one or two open source codebases that claim to enable ecg-based sleep staging. I wonder whether these should be briefly mentioned?

e.g. should comparisons be made to https://github.com/bdsp-core/ecg_respiration_sleep_staging? The repo doesn't look completely baked (e.g. "[training_code_still_messy](https://github.com/bdsp-core/ecg_respiration_sleep_staging/tree/master/training_code_still_messy))", so perhaps it reinforces your point that "either the processing pipeline or the dataset used to generate the published results are not publicly available."
