### 2023-05-19 cbrnr/sleepecg#163 @cbrnr
All (required and optional) dependencies are listed in `pyproject.toml`, and `pip` uses this file for installation. For example, if you want to set up your dev environment, you can do so with `pip install -e ".[dev]"` from the root folder of your clone, which will install all dependencies necessary for development.

I would like to avoid duplicating (and keeping in sync) requirements in a separate `requirements.txt`, unless there is a specific use case that is not covered by `pyproject.toml` – maybe you have something in mind that I'm missing?
