### 2023-05-19 cbrnr/sleepecg#164 @cbrnr
This is a good point. I have been using relative imports in my packages out of habit, with the main reason being that it makes some imports shorter. But you are right, I think it would definitely be worthwile to check if switching to absolute imports would make anything more complex. If not, I think we should go for it. I'll link the corresponding PR so that we can decide if it's worth it or not!
