### 2023-05-19 cbrnr/sleepecg#165 @cbrnr
Personally, I don't like adding boilerplate unless it is absolutely necessary. In my experience, stuff like `if __name__ == "__main__":` and putting code into a `main` function can be very confusing for (new) users, and I would like to keep our examples as simple as possible. I don't think this would impact running the examples from the command line either, but please let me know if I'm missing something.
