### 2023-05-19 cbrnr/sleepecg#166 @cbrnr
I have already addressed this in the tests, but AFAIK the only way to do so is pretty ugly:

```python
try:
    from scipy.datasets import electrocardiogram
except ImportError:
    from scipy.misc import electrocardiogram
```

Which I think would be distracting in the examples. But then again, the `DeprecationWarning` is also distracting, so I'm a bit unsure how to proceed.
