+++ sleepecg/io/__init__.py
@@ -4,10 +4,16 @@
 
 """Functions for downloading and reading datasets."""
 
-from .ecg_readers import ECGRecord, export_ecg_record, read_gudb, read_ltdb, read_mitdb
-from .nsrr import download_nsrr, set_nsrr_token
-from .physionet import download_physionet
-from .sleep_readers import (
+from sleepecg.io.ecg_readers import (
+    ECGRecord,
+    export_ecg_record,
+    read_gudb,
+    read_ltdb,
+    read_mitdb,
+)
+from sleepecg.io.nsrr import download_nsrr, set_nsrr_token
+from sleepecg.io.physionet import download_physionet
+from sleepecg.io.sleep_readers import (
     Gender,
     SleepRecord,
     SleepStage,
