### 2023-05-19 cbb181e
Add missing np import in example (#171)
