### 2023-05-19 cbrnr/sleepecg#168 @cbrnr
I think this is a great suggestion! I assume that by "non-cheat" version you are referring to the fact that our figure shows sleep stages based on the labels in the data, and you would prefer labels predicted by a classifier? What about including the hypnogram from [`examples/try_ws_gru_mesa.py`](https://github.com/cbrnr/sleepecg/blob/main/examples/try_ws_gru_mesa.py)? It has only two stages, so if you prefer a plot with three stages I'd have to adapt the example (shouldn't be too much effort).

![Figure_1](https://github.com/cbrnr/sleepecg/assets/4377312/716a7a64-4b12-42fd-b8ff-ea3d62672226)
