### 2023-05-19 cbrnr/sleepecg#168 @cbrnr
Regarding the metrics figure, I think that it definitely adds value, but I'm not sure how many figures we can and should include in the paper. As part of the review process, I've already added an overview figure, so we currently have two figure (the overview and R peak detector runtimes). I would like to add a hypnogram (to support highlighting the focus of SleepECG) and maybe skip the detector metrics. Let me know what you think (and if you would like a two stage or three stage hypnogram)!
