### 2023-05-19 cbrnr/sleepecg#169 @cbrnr
I agree with both points that we should probably mention related packages and that these are not as readily usable and as open as SleepECG. I don't want to start a discussion on the pros and cons of each package, so maybe briefly mention that there are other packages out there should be sufficient.

Do you have any additional packages that you found other than the one you mentioned?
