### 2023-05-19 cbrnr/sleepecg#157 @richrobe
Alright, that's of course fine for me! Thank you for your efforts 🙂
