### 2023-05-19 cbrnr/sleepecg#157 @cbrnr
The only downside is that it doesn't run on the GPU then. But it should still be plenty fast, and if someone really wants GPU support they can try and use the Apple-provided packages at their own risk.
