### 2023-05-19 3c57a7a
Use official tensorflow package also on macOS (#172)
