### 2023-05-19 cbrnr/sleepecg#156 @cbrnr
@richrobe is the overview figure I've added OK? I didn't really know what to include, so please feel free to comment if you have more ideas on how to improve that figure.

https://raw.githubusercontent.com/openjournals/joss-papers/joss.05411/joss.05411/10.21105.joss.05411.pdf
