### 2023-05-19 #5411 @cbrnr
@tompollard I've addressed most of your comments in the corresponding issues, so here I'm only going to reply to points where this is not the case (but I'll indicate the issues where relevant).

- [ ] Dependencies: See https://github.com/cbrnr/sleepecg/issues/163.
- [x] pip install in zsh: Fixed in https://github.com/cbrnr/sleepecg/pull/160.
- [x] Numba deprecation warning: Fixed in https://github.com/cbrnr/sleepecg/pull/151.
- [x] Relative vs. absolute imports: Fixed in https://github.com/cbrnr/sleepecg/pull/170.
- [ ] Structure in scripts: See https://github.com/cbrnr/sleepecg/issues/165.
- [ ] CLI: Currently I don't see the immediate benefit of adding a CLI, but that's maybe something to keep in mind.
- [ ] `scipy.misc.electrocardiogram`: See https://github.com/cbrnr/sleepecg/issues/166.
- [ ] Figure 1: The y-axis label reads "Mean runtime (s)", which is an average over multiple datasets. I'm not sure it is worth including all the details about how we ran the benchmarks, but I could refer to the description of the benchmark script. I'll also add something like "note the logarithmic scale of the y-axis" to the caption to draw attention to that (important) fact. I'll spell out LTDB. I'll think about how to best describe the different detectors (such as `mne`), because we might be approaching the word limit in this paper. But you are absolutely right that something needs to be added to clarify!
- [ ] Adding more figures: See https://github.com/cbrnr/sleepecg/issues/168.
- [ ] State of field: See https://github.com/cbrnr/sleepecg/issues/169.
- [ ] Reiterate: Agreed, I'll add a sentence near the end of the document!
- [ ] Citing algorithms and data sources: We've cited the Pan & Tompkins algorithm, which other algorithms are you referring to? Currently, we've linked the data sources instead of citing them, because I'm not sure that every database has its own publication (but I could take a look of course).
- [ ] Code archive: A Zenodo DOI will be created as part of the JOSS publication (once and if it gets accepted).
- [x] Documentation: Fixed in https://github.com/cbrnr/sleepecg/pull/171.
