### 2023-05-19 cbrnr/sleepecg#166 @cbrnr
One possible solution could be to wrap this example dataset into our own object (e.g. in `sleepecg.utils`). Thoughts?
