### 2023-05-19 cbrnr/sleepecg#163 @tompollard
@cbrnr you make a good point, there's no particularly strong reason to have a requirement.txt file here (I think it's just become habit to see one!).
