### 2023-05-19 cbrnr/sleepecg#169 @tompollard
> I don't want to start a discussion on the pros and cons of each package, so maybe briefly mention that there are other packages out there should be sufficient.

I agree, just some quick pointers to existing repos would be enough.

> Do you have any additional packages that you found other than the one you mentioned?

Like you say, there doesn't seem to be much around. The repos that I came across during a quick search included:

- https://github.com/Acrophase/Sleep_Staging_KD
- https://github.com/pickus91/ECG-Sleep-Stage-Classification
- https://github.com/ojwalch/sleep_classifiers
- https://github.com/bzhai/multimodal_sleep_stage_benchmark

I'll leave you to decide which of these are worth mentioning, if any!
