### 2023-05-19 cbrnr/sleepecg#166 @tompollard
Hmm, you could keep this issue open and update the dependency and documentation later? Either that or just bump the scipy dependency to 1.10.0 now and get it out of the way!
