### 2023-05-19 cbrnr/sleepecg#168 @tompollard
> I would like to add a hypnogram (to support highlighting the focus of SleepECG) and maybe skip the detector metrics. Let me know what you think (and if you would like a two stage or three stage hypnogram)!

Skipping the metrics makes sense for the reasons you say. I like the idea of a hypnogram. The one you've shown above would be fine I think, perhaps with a note to say that multistage classification is also supported?
