### 2023-05-19 cbrnr/sleepecg#165 @tompollard
> Personally, I don't like adding boilerplate unless it is absolutely necessary. In my experience, stuff like if __name__ == "__main__": and putting code into a main function can be very confusing for (new) users, and I would like to keep our examples as simple as possible.

That's fair, you make a good point, so I'll close this issue.

My feeling is that scripts aren't especially straightforward for beginners, boilerplate or not, because they aren't straightforward to step through, modify, and explore. For this, I think notebooks are the best option, especially if there is an option to open in the cloud.
