### 2023-05-19 cbrnr/sleepecg#165 @cbrnr
Yes, that's true, some basic knowledge of Python, including installing, configuring, and running code is required. Notebooks have their pros and cons, in some situations they are quite useful indeed. I'll keep that in mind, maybe we can add notebooks as examples at a later point.
