### 2023-05-19 cbrnr/sleepecg#166 @cbrnr
Let's keep it open then. I'd like to wait until SciPy 1.11 or 1.12 before I bump it to 1.10. People don't always have access to the latest package versions immediately.
