### 2023-05-22 cbrnr/sleepecg#168 @cbrnr
I've added the hypnogram to the paper.
