### 2023-05-22 cbrnr/sleepecg#156 @richrobe
Sorry - no, the figure gives a nice overview. Maybe play around with the layout and colors a little bit to make it a little more "appealing"... 🙂
