### 2023-05-22 cbrnr/sleepecg#173 @peterakirk
[JOSS Review] HRV Units
Hi all, just a comment relating to the JOSS review (https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1536077814).

I've been testing out your package on on an open dataset (fs=100, https://doi.org/10.13026/sq6q-zg04), just to make sure HRV values look as they should, using the following code (ignoring sleep classification).
```
# sleepecg extraction
sleepecg_beats = sleepecg.detect_heartbeats(sub_ecg.values, fs)
record = sleepecg.SleepRecord(heartbeat_times=sleepecg_beats)
features, stages, feature_ids = sleepecg.extract_features([record], lookback=240, lookforward=60,
                                                          feature_selection=["hrv-time"])
sleepecg_analyzed = features[0][:, 5]*10
sleepecg_index = (sleepecg_analyzed < 100)
sleepecg_rmssd = np.nanmean(sleepecg_analyzed[sleepecg_index])
```

HRV features (i.e., RMSSD) appear to be correlate highly with our (RapidHRV) metrics. However, they appear 1 magnitude smaller in units. For instance, SleepECG suggests subject 80 has an RMSSD of 1.82 which would be highly unusual, where as 18.2 is much more typical and is more in line with what I would expect (RapidHRV puts it close to this too, at 17.83).

When I see RMSSD values, I typically expect them to be in ms units (in the 10-100 range). This doesn’t seem _necessarily_ vital for classification etc., but would be important for interpretation. If I have understood this correctly—that the returned units are not in ms—could you clarify this in your documentation (or change output accordingly, if possible)?
