### 2023-05-22 cbrnr/sleepecg#174 @peterakirk
[JOSS Review] Example usage in readme.md
Just a final, quick comment relating to the JOSS review (https://github.com/openjournals/joss-reviews/issues/5411#issuecomment-1536077814).

You have some nicely laid out, extensive documentation on your readthedocs page. However, when going to a github page, it's nice to just see quick example usage on the repository front page (README.md). Would be great if you could consider including a few lines.
