### 2023-05-22 #5411 @peterakirk
Hi all,

Firstly, well done to @cbrnr and @hofaflo for curating what appears to be a great package. I’m not as familiar with the sleep literature, but classifying sleep stages from ECG is not an approach I had heard of before. This package seems like a great way to make this accessible to the field, in a reproducible and open manner.

I’ve not too much to add. I agree with the other reviewer comments, the most of which seem to have already been addressed by the authors. Below are just a few quick comments (the latter of which I've opened as issues):

- Everything installed smoothly for me (PyCharm 2022.2.2 CE IDE, Python 3.9, Precision 5820, Windows 10, Intel i9).
- I ran a quick sanity check for HRV metrics on an open ECG dataset (https://doi.org/10.13026/sq6q-zg04). RMSSD values appear to correlate highly (ICC=.94, _p_ < .0001) between SleepECG and our package, RapidHRV, after accounting for the lack of unit normalization (see next comment).

![RMSSD_benchmark](https://github.com/openjournals/joss-reviews/assets/50597200/d6073a27-61a4-40d5-9d29-db1d7dad05b6)

- Unless I have misunderstood, HRV metrics (e.g., RMSSD) do not appear to be in normalized units. If I saw RMSSD values, I would typically expect them to be in ms units (in the 10-100 range). This should be explained in the documentation or changed accordingly.
- It would be useful to include a brief example usage on the github readme.md.
