### 2023-05-22 cbrnr/sleepecg#156 @cbrnr
OK, I tried by adding a little color (from the colorbrewer palette), so hopefully it's a bit nicer now.
