### 2023-05-22 5c991c4
Add example to README.md (#175)
