+++ docs/heartbeat_detection.md
@@ -25,16 +25,12 @@ from sleepecg import detect_heartbeats
 ecg = electrocardiogram()  # 5 min of ECG data at 360 Hz
 fs = 360
 beats = detect_heartbeats(ecg, fs)
-print(f"{len(beats)} heartbeats detected ({', '.join(beats[0:5])}, ...)")
-# 478 heartbeats detected (125, 342, 551, 748, 944, ...)
 ```
 
 Next, we calculate RR intervals (in milliseconds):
 
 ```python
 rri = 1000 * np.diff(beats) / fs
-print(np.round(rri[:10]))  # show the first 10 RR intervals
-# array([602, 580, 547, 544, 516, 516, 511, 527, 525, 513])
 ```
 
 Now let's add some random noise to the electrocardiogram and compare the heartbeat detection with and without noise:
@@ -46,10 +42,7 @@ from sleepecg import compare_heartbeats
 np.random.seed(42)
 ecg_noisy = ecg + np.random.rand(ecg.size)
 beats_noisy = detect_heartbeats(ecg_noisy, fs)
-# tp = true positive, fp = false positive, fn = false negative
 tp, fp, fn = compare_heartbeats(beats_noisy, beats)
-print(f"{len(tp)} TP, {len(fp)} FP, {len(fn)} FN")
-# 305 TP, 201 FP, 173 FN
 ```
 
 ## Benchmarks
