### 2023-05-22 #5411 @cbrnr
For some reason, the rendered PDF does not use the latest workflow figure, which should have colored boxes. The correct PDF can be found here: https://github.com/cbrnr/sleepecg/suites/13063524896/artifacts/708420049
