### 2023-05-23 cbrnr/sleepecg#173 @cbrnr
> Yes, absolutely ok to just include in docs. I just think the information needs to be there somewhere, as I was a little confused at first.

OK, I'll update the docs!

> Likewise, it might be useful to also clarify in your docs that sleepecg.ECGRecord fs argument is in ms (not hz).

The sampling frequency should be specified in Hz, not ms. But you are right, this information is not included in the docstring (it just says "The sampling frequency."). I'll add "(in Hz)". Or did you have a specific reason why you thought it was really ms?

> Constraining averaging to windows with an RMSSD of < 100 was just a quick, crude cleaning procedure I used on the data based on a conservative biological constraint, as that is what I applied to benchmark against RapidHRV data too. The expansion by 10 was simply because all the RMSSD values were 10x lower than RMSSD values in ms. However, presumably this is not what you would expect converting RMSSD in seconds to ms? Also, would RMSSD values in seconds be in the range 1-5 like was produced in my analyses? This seems high.

I'm not sure, but since 1s = 1000ms, I'd expect a difference of a factor of 1000, not 10.

> Is this related to the input data sample rate (i.e., fs=100)? I had a quick scan and did not notice resampling included as part of _hrv_timedomain_features (whereas there is for _hrv_frequencydomain_features).

No, this should not be related to the sampling frequency.
