### 2023-05-24 cbrnr/sleepecg#173 @hofaflo
`detect_heartbeats` returns an array of indices, so in the code snippet above, before passing `sleepecg_beats` as `heartbeat_times` to `SleepRecord`, a division by the sampling rate is required to get a factor of 1000.
