### 2023-05-24 d1ee1dc
Clarify that features are based on time in seconds
