+++ .pre-commit-config.yaml
@@ -14,18 +14,10 @@ repos:
     - id: mixed-line-ending
     - id: check-builtin-literals
 
-  - repo: https://github.com/pycqa/flake8
-    rev: 6.0.0
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.269
     hooks:
-    - id: flake8
-
-  - repo: https://github.com/pycqa/pydocstyle
-    rev: 6.3.0
-    hooks:
-    - id: pydocstyle
-      files: ^sleepecg/
-      additional_dependencies:
-      - tomli
+      - id: ruff
 
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.2.0
