+++ .github/workflows/release.yml
@@ -51,8 +51,8 @@ jobs:
 
       - name: Build sdist
         run: |
-          python -m pip install numpy
-          python setup.py sdist
+          python -m pip install build numpy
+          python -m build --sdist
 
       - uses: actions/upload-artifact@v3
         with:
