+++ .github/workflows/release.yml
@@ -60,6 +60,7 @@ jobs:
 
 
   upload_pypi:
+    name: Upload to PyPI
     needs: [pre-commit, build_wheels, build_sdist]
     runs-on: ubuntu-latest
     steps:
