### 2023-06-01 #5411 @cbrnr
The version number to be used in the JOSS paper is [v0.5.5](https://github.com/cbrnr/sleepecg/releases/tag/v0.5.5). The Zenodo DOI is [10.5281/zenodo.7993717](https://zenodo.org/record/7993717).
