### 2023-06-01 #5411 @emdupre
Thank you, @cbrnr !

I've finished reviewing the code and software and am also very happy with the state of the submission. I did have a few editorial requests :

- Overall, please add citations when possible for the described datasets; e.g., MESA, SHHS.
- In the first footnote, the `github.com/bzhai/multimodal_sleep_stage_benchmark` codebase has requested that they be acknowledged with [a citation](https://github.com/bzhai/multimodal_sleep_stage_benchmark#citation). Could you please update to include and cite that reference in the footnote ?
- On L33 and L34, please add abbreviations for MESA and SHHS, as these are subsequently referenced but not explicitly defined.
- On L46, please change to `described in @TFESC:1996 and @Shaffer:2017. ` to correct in-text formatting
- On L52, please add a citation for Tensorflow (I would recommend using the Zenodo record; i.e., https://zenodo.org/record/7987192)
- On references specifically, I noticed that all of your `title` fields are in double-brackets and so following the exact capitalization that you include in [your .bib file](https://github.com/cbrnr/sleepecg/blob/joss/joss/paper.bib). Is there a reason some of these are in sentence case while others are not ? It would be helpful to standardize as possible, either by removing one set of brackets or by explicitly confirming the casing of all titles.
