### 2023-06-07 #5411 @emdupre
Thank you for confirming, @cbrnr !

We're now waiting on JOSS EiC availability ; I will follow-up with that team if we don't hear back in the next couple of days, just to make sure this doesn't slip through the cracks !
