### 2023-06-09 f5883a5
Remove sphinx from dev requirements
