### 2023-10-04 cbrnr/sleepecg#166 @cbrnr
If we take [these support windows](https://scientific-python.org/specs/spec-0000/#support-window) defined by the [Scientific Python project](https://scientific-python.org), we can drop support for SciPy < 1.10 on 28 Jul 2024.
